feat: fix state
